// src/utils/enums/typography.ts
/* eslint-disable no-unused-vars */
/**
 * Enum to define the size of `Paragraph`
 * @enum {string}
 * @readonly
 */
export enum ParagraphView {
  xxSmall = 'xxSmall',
  xSmall = 'xSmall',
  small = 'small',
  medium = 'medium'
}

/**
 * Enum to define the text color of `Paragraph`
 * @enum {string}
 * @readonly
 */
export enum ParagraphColor {
  black = 'black',
  white = 'white',
  gray = 'gray',
  red = 'red',
  green = 'green',
  blue = 'blue'
}

/**
 * Enum to define the size of `Title`
 * @enum {string}
 * @readonly
 */
export enum TitleView {
  small = 'small',
  medium = 'medium',
  large = 'large'
}